Used shared file for vec tests
